Refactor into Streams
Moves much of the logic into Stream instead of in the reducing function.
Bit cleaner (although I think it is certainly a hack to use Keyword like
I am. Doing so because I want to be able to `put_new` and also want to
be able to call `length` on it in a guard...